chore: navbar 문제 해결 중
